feat: abstract input for 2fa into a component
